Fixing issues with Dividend (Basic) functionality (#38)
I ran into a few issues with the Dividend functionality. The DividendFrequency field in the Dividend class was not deserializing correctly because the response from IEXCloud was coming in as "quarterly" (i.e., all lowercase).

I was also seeing an issue when retrieving data for the "Next" dividend because the response from IEXCloud was no longer an array but was an object. I added a new method to the StockFundamentals class to handle this one-off case. I took the same approach with the "Splits" method as well and updated its tests since they were failing since there are no upcoming or prior splits for APPL.

Not sure if these are the most elegant solutions so I am open to feedback on both.

I also removed the EarningsToday method from the StockFundamentals service as it is duplicated in the MarketInfo service. I think that is where it belongs, but let me know if you would prefer to keep it in the StockFundamentals service as well.

Pull request also includes update to address issue #2 which was requesting support for the listLimit query string parameter for the market/list/{listType} endpoint.

Pull request also includes adding of the "Period" parameter to the Estimates method in the StockResearch service.

* Fixing issues with dividend (basic) functionality.

* Removing EarningsToday method from the StockFundamentals service since it is duplicated in the MarketInfo service which I think it where it belongs.

* Adding 'Next' split approach much like what was added for the 'Next' dividend case.

* Adding support for 'listLimit' parameter for market info list endpoint. Addresses issue #2.

* Updating Stock Research Service Estimates call to accept the Period parameter.

* Refactoring the 'Next' case for dividends and splits to use the dynamic keyword instead of adding another method to handle that case.

* Additional refactor for providing standardized interface for Dividend and Split responses when calling with the Next option.